improve structure: split files into groups and multiple files
