+++ .gitignore
@@ -62,3 +62,6 @@ fastlane/report.xml
 fastlane/Preview.html
 fastlane/screenshots
 fastlane/test_output
+
+# Xcode 11 Swift PM
+.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
