+++ .travis.yml
@@ -1,6 +1,6 @@
 
 language: objective-c
-osx_image: xcode10.2
+osx_image: xcode10.3
 script:
     - xcodebuild -enableCodeCoverage=YES clean test -project Endpoints.xcodeproj -scheme Endpoints-macOS -destination "platform=OS X" CODE_SIGN_IDENTITY="" CODE_SIGNING_REQUIRED=NO ONLY_ACTIVE_ARCH=NO
     - bash <(curl -s https://codecov.io/bash) -cF ios
