+++ Sources/Request.swift
@@ -2,131 +2,6 @@ import Foundation
 
 public typealias Parameters = [String: String]
 
-public enum HTTPMethod: String {
-    case get     = "GET"
-    case post    = "POST"
-    case put     = "PUT"
-    case patch   = "PATCH"
-    case delete  = "DELETE"
-    case head    = "HEAD"
-    case options = "OPTIONS"
-    case trace   = "TRACE"
-    case connect = "CONNECT"
-}
-
-/// A type representing an HTTP Body.
-///
-/// A `Body` is used by `Request` to create an `URLRequest`.
-///
-/// Adopted by `Data` and `String`.
-/// - seealso: `FormEncodedBody`, `JSONEncodedBody`.
-public protocol Body {
-    
-    /// Returns HTTP Header parameters required for `self`, if any.
-    ///
-    /// This is usally a "Content-Type" header like "application/json" for a
-    /// JSON encoded Body.
-    ///
-    /// Defaults to `nil`.
-    var header: Parameters? { get }
-    
-    /// The body data that should be sent in an HTTP request.
-    var requestData: Data { get }
-}
-
-extension Body {
-    public var header: Parameters? { return nil }
-}
-
-extension Data: Body {
-    public var requestData: Data { return self }
-}
-
-extension String: Body {
-    public var requestData: Data {
-        guard let data = data(using: .utf8) else {
-            fatalError("cannot convert string to data: \(self)")
-        }
-        return data
-    }
-}
-
-/// A type representing a form encoded HTTP request body.
-public struct FormEncodedBody: Body {
-    
-    /// The parameters dictionary to be encoded.
-    public var parameters: Parameters
-    
-    /// Initialize with a parameters dictionary.
-    /// 
-    /// The parameters do not have to be percent encoded.
-    public init(parameters: Parameters) {
-        self.parameters = parameters
-    }
-    
-    /// Returns "Content-Type": "application/x-www-form-urlencoded".
-    public var header: Parameters? {
-        return [ "Content-Type": "application/x-www-form-urlencoded" ]
-    }
-    
-    /// Returns the encapsulated parameters dictionary as form encoded `Data`.
-    public var requestData: Data {
-        return parameters.map { key, value in
-            return "\(encode(key))=\(encode(value))"
-        }.joined(separator: "&").data(using: .utf8)!
-    }
-    
-    /// add percent encoding to a string suitable for a form encoded request.
-    func encode(_ string: String) -> String {
-        // addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) does not work
-        // see: https://stackoverflow.com/a/8088484
-        // implementation from: https://useyourloaf.com/blog/how-to-percent-encode-a-url-string/
-        let unreserved = "*-._"
-        var allowed = CharacterSet.alphanumerics
-        allowed.insert(charactersIn: unreserved)
-
-        guard let encoded = string.addingPercentEncoding(withAllowedCharacters: allowed) else {
-            fatalError("failed to encode param string: \(string)")
-        }
-        return encoded
-    }
-}
-
-/// A type representing a JSON encoded HTTP request body.
-public struct JSONEncodedBody: Body {
-    public let requestData: Data
-    
-    /// Initialize with a JSON object, an `Array` or a `Dictionary`.
-    ///
-    /// - throws: The error thrown by `JSONSerialization.data(withJSONObject:options:)`, if `jsonObject`
-    /// cannot be encoded.
-    public init(jsonObject: Any) throws {
-        requestData = try JSONSerialization.data(withJSONObject: jsonObject, options: .prettyPrinted)
-    }
-    
-    /// Returns "Content-Type": "application/json".
-    public var header: Parameters? {
-        return [ "Content-Type" : "application/json" ]
-    }
-}
-
-/// A type that can transform itself into an `URLRequest`.
-///
-/// This protocol is adopted by `Request`, `URLRequest`, `URL` and `Call`.
-public protocol URLRequestEncodable: CustomDebugStringConvertible {
-    
-    /// Returns an `URLRequest` configured with the data encapsulated by `self`.
-    var urlRequest: URLRequest { get }
-}
-
-extension URLRequestEncodable {
-    
-    /// Returns the value returned by `cURLRepresentation`.
-    public var debugDescription: String {
-        return cURLRepresentation
-    }
-}
-
 /// A type that encapsulates all data necessary to construct an `URLRequest`
 /// suitable for a typical REST-API call.
 ///
@@ -162,17 +37,6 @@ public struct Request: URLRequestEncodable {
         urlRequest.apply(header: header) //request header trumps body header
         
         return urlRequest
-    }
-}
-
-public extension URLRequest {
-    
-    /// Adds or replaces all header fields with the given values.
-    ///
-    /// - note: If a value was previously set for the given header
-    /// field, that value is replaced.
-    mutating func apply(header: Parameters?) {
-        header?.forEach { setValue($1, forHTTPHeaderField: $0) }
     }
 }
