+++ Sources/Result.swift
@@ -1,0 +1,9 @@
+//
+//  Result.swift
+//  Endpoints
+//
+//  Created by Robin Mayerhofer on 27.08.19.
+//  Copyright © 2019 Tailored Apps. All rights reserved.
+//
+
+import Foundation
